Move link server to an included build
