+++ build.gradle.kts
@@ -3,8 +3,6 @@ import nl.littlerobots.vcu.plugin.resolver.VersionSelectors
 plugins {
     `kotlin-conventions`
 
-    alias(libs.plugins.kotlinx.serialization)
-
     alias(libs.plugins.version.catalog.update)
 }
 
@@ -35,7 +33,6 @@ dependencies {
     implementation(libs.kotlin.logging)
 
     // JSON
-    implementation(libs.kotlinx.serialization.json)
     implementation(libs.gson)
 
     // SQL
@@ -43,21 +40,6 @@ dependencies {
     implementation(libs.flyway.core)
     runtimeOnly(libs.flyway.postgresql)
     implementation(libs.hikaricp)
-
-    // Web server
-    implementation(libs.ktor.server.core)
-    implementation(libs.ktor.server.netty)
-    implementation(libs.ktor.server.call.logging)
-    implementation(libs.ktor.server.content.negotiation)
-    implementation(libs.ktor.serialization.kotlinx.json)
-    implementation(libs.ktor.server.resources)
-
-    // Kotlin metadata
-    implementation(libs.kotlin.metadata)
-
-    // JUnit
-    testImplementation(libs.junit.jupiter)
-    testRuntimeOnly("org.junit.platform:junit-platform-launcher")
 }
 
 kotlin.compilerOptions {
