+++ link-server/settings.gradle.kts
@@ -1,0 +1,9 @@
+rootProject.name = "link-server"
+
+dependencyResolutionManagement {
+    versionCatalogs {
+        create("libs") {
+            from(files("../gradle/libs.versions.toml"))
+        }
+    }
+}
