+++ link-server/src/main/kotlin/dev/freya02/link/server/LinkException.kt
@@ -1,0 +1,3 @@
+package dev.freya02.link.server
+
+class LinkException(message: String) : IllegalArgumentException(message)
