+++ link-server/src/main/kotlin/dev/freya02/link/server/LinkServer.kt
@@ -1,10 +1,8 @@
-package io.github.freya022.link.server
+package dev.freya02.link.server
 
-import ch.qos.logback.classic.ClassicConstants
-import io.github.freya022.link.server.resolution.ClassMemberResolver
-import io.github.freya022.link.server.resolution.ClassResolver
-import io.github.freya022.link.server.resolution.TopLevelResolver
-import io.github.freya022.wiki.config.Environment
+import dev.freya02.link.server.resolution.ClassMemberResolver
+import dev.freya02.link.server.resolution.ClassResolver
+import dev.freya02.link.server.resolution.TopLevelResolver
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.ktor.http.*
 import io.ktor.resources.*
@@ -19,34 +17,29 @@ import io.ktor.server.resources.Resources
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import org.slf4j.event.Level
-import kotlin.io.path.absolutePathString
 
 private val logger by lazy { KotlinLogging.logger { } }
 
 @Resource("/link")
 class Link(val identifier: String)
 
-fun main() {
-    System.setProperty(ClassicConstants.CONFIG_FILE_PROPERTY, Environment.logbackConfigPath.absolutePathString())
-    logger.info { "Loading logback configuration at ${Environment.logbackConfigPath.absolutePathString()}" }
+object LinkServer {
+    /** Needs to be started */
+    fun embeddedLinkServer() =
+        embeddedServer(Netty, port = 16069) {
+            install(ContentNegotiation) {
+                json()
+            }
+            install(Resources)
+            install(CallLogging) {
+                level = Level.DEBUG
+            }
 
-    embeddedLinkServer().start(wait = true)
+            routing {
+                link()
+            }
+        }
 }
-
-fun embeddedLinkServer() =
-    embeddedServer(Netty, port = 16069) {
-        install(ContentNegotiation) {
-            json()
-        }
-        install(Resources)
-        install(CallLogging) {
-            level = Level.DEBUG
-        }
-
-        routing {
-            link()
-        }
-    }
 
 private fun Routing.link() {
     get<Link> {
