+++ link-server/src/main/kotlin/dev/freya02/link/server/utils/ClassPathUtils.kt
@@ -1,4 +1,4 @@
-package io.github.freya022.link.server.utils
+package dev.freya02.link.server.utils
 
 import io.github.classgraph.ClassGraph
 import io.github.freya022.botcommands.api.core.BotCommands
