+++ link-server/src/main/kotlin/dev/freya02/link/server/utils/KotlinMember.kt
@@ -1,4 +1,4 @@
-package io.github.freya022.link.server.utils
+package dev.freya02.link.server.utils
 
 import kotlin.metadata.KmFunction
 import kotlin.metadata.KmProperty
