--- src/main/kotlin/io/github/freya022/link/server/LinkException.kt
@@ -1,3 +1,0 @@
-package io.github.freya022.link.server
-
-class LinkException(message: String) : IllegalArgumentException(message)
